+++ universal_realtime_stt_tts/stt_provider_google.py
@@ -7,8 +7,8 @@ from typing import AsyncIterator, Optional
 
 from google.cloud import speech
 
-from universal_realtime_audio._event_queue import SttEventQueue
-from universal_realtime_audio.stt_provider import RealtimeSttProvider, TranscriptEvent
+from universal_realtime_stt_tts._event_queue import SttEventQueue
+from universal_realtime_stt_tts.stt_provider import RealtimeSttProvider, TranscriptEvent
 
 logger = getLogger(__name__)